dotenv and getBuild in separate files
