+++ packages/back-end/src/init/dotenv.ts
@@ -1,0 +1,6 @@
+import fs from "fs";
+import dotenv from "dotenv";
+
+if (fs.existsSync(".env.local")) {
+  dotenv.config({ path: ".env.local" });
+}
