+++ packages/back-end/src/instrumentation.ts
@@ -1,5 +1,6 @@
+// Warning: Careful importing other modules, as they and any dependencies they import won't be instrumented.
 import * as Sentry from "@sentry/node";
-import { getBuild } from "./util/handler";
+import { getBuild } from "./util/build";
 
 const SENTRY_DSN = process.env.SENTRY_DSN;
 if (SENTRY_DSN) {
