+++ packages/back-end/src/server.ts
@@ -1,5 +1,5 @@
 import "./init/aliases";
-import "./util/secrets";
+import "./init/dotenv";
 import "./instrumentation";
 import app from "./app";
 import { logger } from "./util/logger";
