+++ packages/back-end/src/util/handler.ts
@@ -1,5 +1,3 @@
-import path from "path";
-import fs from "fs";
 import { Request, RequestHandler } from "express";
 import z, { Schema, ZodNever } from "zod";
 import { orgHasPremiumFeature } from "back-end/src/enterprise";
@@ -127,47 +125,6 @@ export function createApiRequestHandler<
     };
     return wrappedHandler;
   };
-}
-
-let build: { sha: string; date: string; lastVersion: string };
-export function getBuild() {
-  if (!build) {
-    build = {
-      sha: "",
-      date: "",
-      lastVersion: "",
-    };
-    const rootPath = path.join(__dirname, "..", "..", "..", "..", "buildinfo");
-    if (fs.existsSync(path.join(rootPath, "SHA"))) {
-      build.sha = fs.readFileSync(path.join(rootPath, "SHA")).toString().trim();
-    }
-    if (fs.existsSync(path.join(rootPath, "DATE"))) {
-      build.date = fs
-        .readFileSync(path.join(rootPath, "DATE"))
-        .toString()
-        .trim();
-    }
-
-    // Read version from package.json
-    try {
-      const packageJSONPath = path.join(
-        __dirname,
-        "..",
-        "..",
-        "..",
-        "..",
-        "package.json"
-      );
-      if (fs.existsSync(packageJSONPath)) {
-        const json = JSON.parse(fs.readFileSync(packageJSONPath).toString());
-        build.lastVersion = json.version;
-      }
-    } catch (e) {
-      // Ignore errors here, not important
-    }
-  }
-
-  return build;
 }
 
 export async function validateIsSuperUserRequest(req: {
